Fix schema validator
Schema validator would throw an error if value given for type_ is not a type,
and never get to the fallback check. This fix will first check if the given
type is an actual type, and then do the actual `isinstance` check. If
the validator was not given an actual type, then the value is assumed
to pass validation.