Fixed how retirement happens
If there aren't a lot of gold labels for a class of labels, then the thresholds
for retirement will be really far from the appropriate label. If the thresholds
surpass the prior, then every subject would be immediately retired as that
label. Ex if the prior is .12 and the bogus threshold is ~.5, then every subject
is immediately retired as bogus.

If a threshold has passed the prior, that threshold is set to the prior. This
commit also changes what is stored where for retirement, so that the retired
label can then be pulled directly from the score object.